requirements.old.txt
BeautifulSoup==3.2.1
ConfigArgParse==0.11.0
Pillow==2.3.0
beautifulsoup4==4.5.3
yandex.translate==0.3.5
ipython==1.2.1
mock==1.0.1
numpy==1.8.2
parsedatetime==2.1
poster==0.8.1
pyTelegramBotAPI==2.3.1
pygeoip==0.3.2
pyowm==2.7.1
requests==2.12.4
simplegeneric==0.8.1
socketIO-client==0.5.6
urllib3==1.19.1
websocket-client==0.40.0
wheel==0.24.0
gTTS==1.2.2
tweepy==3.5.0
PyDictionary==1.5.2
